Pedro prefers Inch measurements. Also move tests to tests folder
