+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/LimelightAimingTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.testing;
 
 import com.qualcomm.hardware.limelightvision.LLResult;
 import com.qualcomm.hardware.limelightvision.LLResultTypes.FiducialResult;
@@ -13,7 +13,7 @@ import java.util.List;
 // Note: Ensure your Limelight 3a is configured via its web interface to use the AprilTag pipeline
 // and is named "limelight" in your robot's hardware configuration.
 
-@TeleOp(name = "Limelight AprilTag Test", group = "Test")
+@TeleOp(name = "Limelight AprilTag Test", group = "Tests")
 public class LimelightAimingTest extends LinearOpMode {
 	
 	// Define the specific AprilTag ID we are looking for.
